Modified the VM's opcode internal function to take the number of seconds to use for the SLEEP instruction from a register, which is the subsequent byte in the program, interpreted as the register to use.
In notional "machine code", this instruction will look like this:

   SLEEP reg3

which would imply to "sleep a number of seconds based on the integer that
is stored at register 3". Obviously a well-written program built by a high
level language compiler will store the integer value in that register, and
so it will look something like this:

   INT_STORE reg3 60
   SLEEP reg3

If found anywhere inside a program, these two instructions will set the
number "60" inside register "3", and then the sleep()ing instruction will
be called to use the contents of this register.

(The use of a register with the SLEEP instruction instead of a literal
number in a program makes more sense because a program will expect that, in
general, the sleep function will take use a value that is a result of some
calculation, and therefore the result of prior operations that must be
stored in memory, and therefore it will be fetched and stored in a register
for immediate use. This is based on my understanding of how things should
work.)
IN 2021/11/12